+++ app/database/models/project_works.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 from datetime import datetime
 from sqlalchemy.sql import text
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, UUID, ForeignKey, Boolean, Numeric, Integer
+from sqlalchemy import Column, UUID, ForeignKey, Boolean, Numeric, BigInteger
 from app.database.db_setup import Base
 
 
@@ -16,7 +16,7 @@ class ProjectWorks(Base):
     quantity = Column(Numeric(precision=10, scale=2), nullable=False)
     summ = Column(Numeric(precision=10, scale=2), nullable=True)
     signed = Column(Boolean, nullable=False, default=False)
-    created_at = Column(Integer, default=lambda: int(datetime.utcnow().timestamp()),
+    created_at = Column(BigInteger, default=lambda: int(datetime.utcnow().timestamp()),
                         server_default=text("EXTRACT(EPOCH FROM NOW())"), nullable=False)
     created_by = Column(UUID, ForeignKey(
         'users.user_id'), nullable=False)
