+++ app/database/models/work_prices.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 from datetime import datetime
 from sqlalchemy.sql import text
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, Integer, UUID, ForeignKey, Boolean, Numeric, CheckConstraint
+from sqlalchemy import Column, Integer, UUID, ForeignKey, Boolean, Numeric, CheckConstraint, BigInteger
 from app.database.db_setup import Base
 
 
@@ -14,7 +14,7 @@ class WorkPrices(Base):
     work = Column(UUID, ForeignKey('works.work_id'))
     category = Column(Integer, nullable=False)
     price = Column(Numeric(precision=10, scale=2), nullable=False)
-    created_at = Column(Integer, default=lambda: int(datetime.utcnow().timestamp()),
+    created_at = Column(BigInteger, default=lambda: int(datetime.utcnow().timestamp()),
                         server_default=text("EXTRACT(EPOCH FROM NOW())"), nullable=False)
     created_by = Column(UUID, ForeignKey(
         'users.user_id'), nullable=False)
